+++ packages/components/src/components/Navigation/components/NavigationGroup/NavigationGroup.tsx
@@ -65,7 +65,6 @@ export const NavigationGroup = flowComponent("NavigationGroup", (props) => {
           {...rest}
         >
           {children}
-          <TunnelExit id="Label" />
           <ul>
             <TunnelExit id="groupLinks" />
           </ul>