+++ packages/components/src/components/NotificationProvider/NotificationContainer/NotificationContainer.tsx
@@ -4,7 +4,13 @@ import ReactDOM from "react-dom";
 import clsx from "clsx";
 import styles from "./NotificationContainer.module.scss";
 import ControlledNotification from "@/components/NotificationProvider/ControlledNotification";
-import { AnimatePresence, domAnimation, LazyMotion, m } from "framer-motion";
+import {
+  AnimatePresence,
+  domAnimation,
+  LazyMotion,
+  m,
+  MotionConfig,
+} from "framer-motion";
 import { useIsSSR } from "react-aria";
 import { useNotificationController } from "@/components/NotificationProvider/NotificationProvider";
 
@@ -22,31 +28,33 @@ export const NotificationContainer: FC<NotificationsContainerProps> = (
   const rootClassName = clsx(styles.notificationContainer, className);
 
   const content = (
-    <LazyMotion features={domAnimation}>
-      <div className={rootClassName} {...rest}>
-        <AnimatePresence>
-          {notifications.map((n) => (
-            <m.div
-              className={styles.notification}
-              key={n.meta.id}
-              initial={{ opacity: 0, x: 200 }}
-              animate={{ opacity: 1, x: 0 }}
-              exit={{ opacity: 0, x: 200, height: 0, paddingBottom: 0 }}
-              transition={{
-                bounce: 0,
-              }}
-            >
-              <Suspense>
-                <ControlledNotification
-                  notification={n}
-                  controller={controller}
-                />
-              </Suspense>
-            </m.div>
-          ))}
-        </AnimatePresence>
-      </div>
-    </LazyMotion>
+    <MotionConfig reducedMotion="user">
+      <LazyMotion features={domAnimation}>
+        <div className={rootClassName} {...rest}>
+          <AnimatePresence>
+            {notifications.map((n) => (
+              <m.div
+                className={styles.notification}
+                key={n.meta.id}
+                initial={{ opacity: 0, x: 200 }}
+                animate={{ opacity: 1, x: 0 }}
+                exit={{ opacity: 0, x: 200, height: 0, paddingBottom: 0 }}
+                transition={{
+                  bounce: 0,
+                }}
+              >
+                <Suspense>
+                  <ControlledNotification
+                    notification={n}
+                    controller={controller}
+                  />
+                </Suspense>
+              </m.div>
+            ))}
+          </AnimatePresence>
+        </div>
+      </LazyMotion>
+    </MotionConfig>
   );
 
   return isSsr ? null : ReactDOM.createPortal(content, document.body);