+++ packages/components/src/lib/hooks/useFieldError.tsx
@@ -3,7 +3,7 @@ import { type PropsContext, PropsContextProvider } from "@/lib/propsContext";
 import formFieldStyles from "@/components/FormField/FormField.module.scss";
 import { TunnelExit } from "@mittwald/react-tunnel";
 import ClearPropsContext from "@/lib/propsContext/components/ClearPropsContext";
-import { useProps } from "@/index/default";
+import { useProps } from "@/lib/hooks/useProps";
 
 export const useFieldError = (tunnelIdFromProps?: string) => {
   const id = useId();