+++ packages/remote-react-components/src/tests/visual/SkeletonText.browser.test.tsx
@@ -1,0 +1,16 @@
+import { testEnvironments } from "@/tests/lib/environments";
+import { test } from "vitest";
+
+test.each(testEnvironments)(
+  "SkeletonText (%s)",
+  async ({ testScreenshot, render, components: { SkeletonText, Flex } }) => {
+    await render(
+      <Flex direction="column" gap="m">
+        <SkeletonText />
+        <SkeletonText width="300px" />
+      </Flex>,
+    );
+
+    await testScreenshot("SkeletonText");
+  },
+);